Remove duplicate pytest fixtures, promote mos fixtures to global conftest.py
