+++ jdaviz/configs/mosviz/tests/test_helper.py
@@ -1,128 +1,19 @@
-import warnings
-import astropy.units as u
+# Tests all non-data-loading functionality of the Mosviz Helper class
+# For data-loading tests, see test_data_loading.py
+
 import csv
-import numpy as np
 import pytest
 from numpy.testing import assert_allclose
-from astropy.nddata import CCDData
-from astropy.wcs import WCS
-from jdaviz.configs.mosviz.helper import Mosviz
 from jdaviz.configs.specviz2d.helper import Specviz2d
-from specutils import Spectrum1D, SpectrumCollection
-
-
-@pytest.fixture
-def mosviz_helper():
-    return Mosviz()
-
-
-@pytest.fixture
-def spectrum1d():
-    spec_axis = np.linspace(6000, 8000, 1024) * u.AA
-    flux = (np.random.randn(len(spec_axis.value)) +
-            10*np.exp(-0.001*(spec_axis.value-6563)**2) +
-            spec_axis.value/500) * u.Jy
-
-    return Spectrum1D(spectral_axis=spec_axis, flux=flux)
-
-
-@pytest.fixture
-def spectrum2d():
-    header = """
-WCSAXES =                    2 / Number of coordinate axes
-CRPIX1  =                  0.0 / Pixel coordinate of reference point
-CRPIX2  =               1024.5 / Pixel coordinate of reference point
-CDELT1  =                1E-06 / [m] Coordinate increment at reference point
-CDELT2  =  2.9256727777778E-05 / [deg] Coordinate increment at reference point
-CUNIT1  = 'm'                  / Units of coordinate increment and value
-CUNIT2  = 'deg'                / Units of coordinate increment and value
-CTYPE1  = 'WAVE'               / Vacuum wavelength (linear)
-CTYPE2  = 'OFFSET'             / Spatial offset
-CRVAL1  =                  0.0 / [m] Coordinate value at reference point
-CRVAL2  =                  5.0 / [deg] Coordinate value at reference point
-RADESYS = 'ICRS'               / Equatorial coordinate system
-SPECSYS = 'BARYCENT'           / Reference frame of spectral coordinates
-"""
-    new_hdr = {}
-
-    for line in header.split('\n'):
-        try:
-            key, value = line.split('=')
-            key = key.strip()
-            value, _ = value.split('/')
-            value = value.strip()
-            value = value.strip("'")
-        except ValueError:
-            continue
-
-        new_hdr[key] = value
-
-    wcs = WCS(new_hdr)
-    data = np.random.sample((1024, 15)) * u.one
-    spectral_cube = Spectrum1D(data, wcs=wcs)
-
-    return spectral_cube
-
-
-@pytest.fixture
-def spectrum_collection(spectrum1d):
-    sc = [spectrum1d]*5
-
-    with warnings.catch_warnings():
-        warnings.simplefilter('ignore')
-        result = SpectrumCollection.from_spectra(sc)
-    return result
-
-
-@pytest.fixture
-def image():
-    header = """
-WCSAXES =                    2 / Number of coordinate axes
-CRPIX1  =                937.0 / Pixel coordinate of reference point
-CRPIX2  =                696.0 / Pixel coordinate of reference point
-CDELT1  = -1.5182221158397E-05 / [deg] Coordinate increment at reference point
-CDELT2  =  1.5182221158397E-05 / [deg] Coordinate increment at reference point
-CUNIT1  = 'deg'                / Units of coordinate increment and value
-CUNIT2  = 'deg'                / Units of coordinate increment and value
-CTYPE1  = 'RA---TAN'           / Right ascension, gnomonic projection
-CTYPE2  = 'DEC--TAN'           / Declination, gnomonic projection
-CRVAL1  =      5.0155198140981 / [deg] Coordinate value at reference point
-CRVAL2  =       5.002450989248 / [deg] Coordinate value at reference point
-LONPOLE =                180.0 / [deg] Native longitude of celestial pole
-LATPOLE =       5.002450989248 / [deg] Native latitude of celestial pole
-DATEREF = '1858-11-17'         / ISO-8601 fiducial time
-MJDREFI =                  0.0 / [d] MJD of fiducial time, integer part
-MJDREFF =                  0.0 / [d] MJD of fiducial time, fractional part
-RADESYS = 'ICRS'               / Equatorial coordinate system
-"""
-    new_hdr = {}
-
-    for line in header.split('\n'):
-        try:
-            key, value = line.split('=')
-            key = key.strip()
-            value, _ = value.split('/')
-            value = value.strip()
-            value = value.strip("'")
-        except ValueError:
-            continue
-
-        new_hdr[key] = value
-
-    wcs = WCS(new_hdr)
-    data = np.random.sample((55, 55))
-    ccd_data = CCDData(data, wcs=wcs, unit='Jy')
-
-    return ccd_data
 
 
 @pytest.mark.filterwarnings('ignore')
-def test_viewer_axis_link(mosviz_helper, spectrum1d, spectrum2d):
+def test_viewer_axis_link(mosviz_helper, mos_spectrum1d, mos_spectrum2d):
     label1d = "Test 1D Spectrum"
-    mosviz_helper.load_1d_spectra(spectrum1d, data_labels=label1d)
+    mosviz_helper.load_1d_spectra(mos_spectrum1d, data_labels=label1d)
 
     label2d = "Test 2D Spectrum"
-    mosviz_helper.load_2d_spectra(spectrum2d, data_labels=label2d)
+    mosviz_helper.load_2d_spectra(mos_spectrum2d, data_labels=label2d)
 
     table = mosviz_helper.app.get_viewer('table-viewer')
     table.widget_table.vue_on_row_clicked(0)
@@ -131,7 +22,7 @@ def test_viewer_axis_link(mosviz_helper, spectrum1d, spectrum2d):
     scale_1d = mosviz_helper.app.get_viewer('spectrum-viewer').scales['x']
 
     scale_2d.min = 200.0
-    assert scale_1d.min == spectrum1d.spectral_axis.value[200]
+    assert scale_1d.min == mos_spectrum1d.spectral_axis.value[200]
 
     scale_1d.max = 7564
     assert scale_2d.max == 800.0
@@ -159,11 +50,11 @@ def test_to_csv(tmp_path, mosviz_helper, spectrum_collection):
 
 
 @pytest.mark.filterwarnings('ignore')
-def test_table_scrolling(mosviz_helper, image, spectrum1d, spectrum2d):
+def test_table_scrolling(mosviz_helper, mos_image, spectrum1d, mos_spectrum2d):
     spectra1d = [spectrum1d] * 2
-    spectra2d = [spectrum2d] * 2
+    spectra2d = [mos_spectrum2d] * 2
 
-    mosviz_helper.load_data(spectra1d, spectra2d, images=image)
+    mosviz_helper.load_data(spectra1d, spectra2d, images=mos_image)
 
     table = mosviz_helper.app.get_viewer('table-viewer')
     # first row is automatically selected in the UI
@@ -179,11 +70,11 @@ def test_table_scrolling(mosviz_helper, image, spectrum1d, spectrum2d):
 
 
 @pytest.mark.filterwarnings('ignore')
-def test_column_visibility(mosviz_helper, image, spectrum1d, spectrum2d):
+def test_column_visibility(mosviz_helper, mos_image, spectrum1d, mos_spectrum2d):
     spectra1d = [spectrum1d] * 2
-    spectra2d = [spectrum2d] * 2
+    spectra2d = [mos_spectrum2d] * 2
 
-    mosviz_helper.load_data(spectra1d, spectra2d, images=image)
+    mosviz_helper.load_data(spectra1d, spectra2d, images=mos_image)
 
     with pytest.raises(
                 ValueError,
@@ -202,11 +93,11 @@ def test_column_visibility(mosviz_helper, image, spectrum1d, spectrum2d):
 
 
 @pytest.mark.filterwarnings('ignore')
-def test_custom_columns(mosviz_helper, image, spectrum1d, spectrum2d):
+def test_custom_columns(mosviz_helper, mos_image, spectrum1d, mos_spectrum2d):
     spectra1d = [spectrum1d] * 2
-    spectra2d = [spectrum2d] * 2
+    spectra2d = [mos_spectrum2d] * 2
 
-    mosviz_helper.load_data(spectra1d, spectra2d, images=image)
+    mosviz_helper.load_data(spectra1d, spectra2d, images=mos_image)
 
     mosviz_helper.add_column('custom_name')
     assert 'custom_name' in mosviz_helper.get_column_names(True)
@@ -232,11 +123,11 @@ def test_custom_columns(mosviz_helper, image, spectrum1d, spectrum2d):
 
 
 @pytest.mark.filterwarnings('ignore')
-def test_redshift_column(mosviz_helper, image, spectrum1d, spectrum2d):
+def test_redshift_column(mosviz_helper, mos_image, spectrum1d, mos_spectrum2d):
     spectra1d = [spectrum1d] * 2
-    spectra2d = [spectrum2d] * 2
+    spectra2d = [mos_spectrum2d] * 2
 
-    mosviz_helper.load_data(spectra1d, spectra2d, images=image)
+    mosviz_helper.load_data(spectra1d, spectra2d, images=mos_image)
 
     mosviz_helper.update_column("Redshift", 0.1, row=0)
     assert_allclose(list(mosviz_helper.specviz.get_spectra().values())[0].redshift.value, 0.1)
